requirements.txt
phidata
groq
python-dotenv
openai
streamlit